+++ PaymeFeature/PresentationLayer/MainScreen/FamilyScreen/Views/TransferSheetsSheets/TransactionSheet.swift
@@ -31,7 +31,7 @@ struct TransactionSheet: View {
     }
     
     var isAmountValid: Bool {
-        if let amount = Double(sum), amount > 0 {
+        if let amount = Int(sum), amount > 0 {
             return true
         }
         return false
